requirements.txt
keras==2.10.0
matplotlib==3.6.0
numpy==1.23.3
opencv_python==4.6.0.66
pandas==1.5.0
Pillow==9.2.0
scikit_learn==1.1.2
seaborn==0.12.1
streamlit==1.13.0
